Refactor: using CreateTemplateCommand instead of separate command for initializing the template
