--- src/main/java/edu/harvard/iq/dataverse/api/dto/TermsOfUseAndAccessDTO.java
@@ -1,2 +1,0 @@
-package edu.harvard.iq.dataverse.api.dto;public class TermsOfUseAndAccessDTO {
-}
